Made to compile on a Windows 10 machine

+++ src/uploadwidget.cpp
@@ -23,6 +23,6 @@ void UploadWidget::SetMessage( QString const & message )
 
 void UploadWidget::SetProgress( qint64 current_value, qint64 total )
 {
-    ui->data_item_progress->setMaximum( total + ( total * 0.25 ) );
-    ui->data_item_progress->setValue( current_value );
+    ui->data_item_progress->setMaximum( static_cast<int> ( total + ( total * 0.25 )) );
+    ui->data_item_progress->setValue( static_cast<int>( current_value ) );
 }
